OHMYGODSITWORKED:calling the member function 'scc' in a graph will return the number of connected components.
'finishorder' takes a dfs ordering and strips it down to the finish times.

modified 'dfs' with a boolean argument to either search the whole graph or just the root.

used dictionaries for dynamic purposes, mainly because i'm using the size of the graph for certain measurements, so it's easier to delete entries.

cried my little eyes out.